% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{data}
\name{data_nw}
\alias{data_nw}
\title{A simulated data set for testing}
\format{
An object of class \code{matrix} with 12 rows and 12 columns.
}
\usage{
data_nw
}
\description{
A simulated data set for testing
}
\keyword{datasets}
