% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{data}
\name{data_th}
\alias{data_th}
\title{A simulated data set for testing}
\format{
An object of class \code{list} of length 8.
}
\usage{
data_th
}
\description{
A simulated data set for testing
}
\keyword{datasets}
