% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{data}
\name{data_w}
\alias{data_w}
\title{A simulated data set for testing}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 50 rows and 50 columns.
}
\usage{
data_w
}
\description{
A simulated data set for testing
}
\keyword{datasets}
