% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msdpdth.R
\name{msdpdth}
\alias{msdpdth}
\title{M-estimator for threshold spatial dynamic panel data model}
\usage{
msdpdth(
  y,
  x,
  w1,
  th,
  correction = T,
  max_try = 5,
  all_er = F,
  true_range = F,
  residual = F,
  w3 = w1,
  w2 = w1,
  no_tf = F,
  model = "full",
  rcpp = T,
  cma_pop_multi = 1
)
}
\arguments{
\item{y}{matrix, containing regional index (first column), time index (second column, numeric) and dependent variable (third column, numeric).}

\item{x}{matrix, containing regional index (first column), time index (second column, numeric) and regressors (numeric).}

\item{w1}{matrix, the spatial weight matrix. If w2 and w3 are supplied, the spatial weight matrix for spatial lag.}

\item{th}{logical,}

\item{correction}{logical, whether to use adjusted score function. Default value is True.}

\item{max_try}{integer, maximum attempt for the solver. Default value is 5.}

\item{all_er}{logical, whether to output Hessian and Gamma matrix based se. Ignored if correction is set to False. Default value is False.}

\item{true_range}{logical, whether to used the accurate stationary check. Default value is False due to performance reasons.}

\item{residual}{logical, whether to output the residual. Default value is False.}

\item{w3}{matrix, the spatial weight matrix for spatial error. Default value is the same as w1.}

\item{w2}{matrix, the spatial weight matrix for spatio-temporal lag. Default value is the same as w1.}

\item{no_tf}{logical, whether to account for time effect. Default value is True.}

\item{model}{character, indicates the model used for estimation, can be "full", "slm", "sem", "sltl". See Details.}

\item{rcpp}{logical, whether to use the rcpp implementation to calculate the score function. Default value is True.}

\item{cma_pop_multi}{integer, multiplier for the population size used in cmaes. Default value is 1.}
}
\value{
A list of estimation results
\itemize{
\item{"coefficient"} {list, coefficients and standard errors}
\item{"vc_mat"} {matrix , variance-covariance matrix}
\item{"hes_mat"} {matrix, Hessian matrix}
\item{"gamma_mat"} {matrix, Gamma matrix}
\item{"model"} {character, model used for estimation}
\item{"residual"} {numeric, residual}
}
}
\description{
M-estimator for threshold spatial dynamic panel data model
}
\details{
Estimating threshold spatial dynamic panel data model with extended Yang(2018)'s M-estimator
\deqn{y_{ti} = \mu_{i} +\alpha_t+ x_{ti}\beta_{q} +\rho_{q} y_{t-1,i} + \lambda_{1q}\sum_{j=1}^{n}w_{1,ij}y_{tj} \\ 
\qquad + \lambda_{2q}\sum_{j=1}^{n}w_{2,ij}y_{t-1,i}+ u_{ti},\\
u_{ti} = \lambda_{3q}\sum_{j=1}^{n}w_{3,ij}u_{tj}+ v_{ti},i=1,\ldots,n,t=1,\ldots,T, q = 1,2}
Sub-models can be estimated by changing argument "model"
\itemize{
\item{"full"} {Full model}
\item{"slm"} {\eqn{\lambda_{2q} = \lambda_{3q} = 0}}
\item{"sem"} {\eqn{\lambda_{1q} = \lambda_{2q} = 0}}
\item{"sltl"} {\eqn{\lambda_{3q} = 0}}
}
}
\examples{
data(data_th, data_w)
msdpdth(y = data_th$y, x = data_th$x, w1 = data_w, th = data_th$th)

}
\references{
Yang, Z. (2018). Unified M-estimation of fixed-effects spatial dynamic models with short panels. Journal of Econometrics, 205(2), 423-447.
}
