% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.msdpd}
\alias{print.msdpd}
\title{Print method for msdpd class}
\usage{
\method{print}{msdpd}(x, ...)
}
\arguments{
\item{x}{msdpd class}

\item{...}{other parameters}
}
\value{
A data.frame containing the coefficients and the corresponding standard error.
}
\description{
Print method for msdpd class
}
\details{
Print method for msdpd class
}
\examples{
\donttest{
data(data_n, data_nw)
result <- msdpd(y = data_n$y, x = data_n$x, w1 = data_nw)
result
}
}
