% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.msdpdth}
\alias{print.msdpdth}
\title{Print method for msdpdth class}
\usage{
\method{print}{msdpdth}(x, ...)
}
\arguments{
\item{x}{msdpdth class}

\item{...}{other parameters}
}
\description{
Print method for msdpdth class
}
\details{
Print method for msdpdth class
}
