% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wald_test.R
\name{wald_test}
\alias{wald_test}
\title{Wald test for threshold spatial dynamic panel data model}
\usage{
wald_test(th_res)
}
\arguments{
\item{th_res}{msdpdth class, output from msdpdth}
}
\value{
A list of p-values for each parameters.
}
\description{
Wald test for threshold spatial dynamic panel data model
}
\details{
Two sided Wald test for testing whether two parameters for each groups are equal
\itemize{
\item{"h_0"} {\eqn{\theta_1 = \theta_2}}
\item{"h_1"} {\eqn{\theta_1 \neq \theta_2}}
}
}
