% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.msdpdth}
\alias{print.msdpdth}
\title{Print method for msdpdth class}
\usage{
\method{print}{msdpdth}(x, ...)
}
\arguments{
\item{x}{msdpdth class}

\item{...}{other parameters}
}
\value{
A data.frame containing the coefficients and the corresponding standard error.
}
\description{
Print method for msdpdth class
}
\details{
Print method for msdpdth class
}
\examples{
\donttest{
data(data_th, data_w)
result <- msdpdth(y = data_th$y, x = data_th$x, w1 = data_w, th = data_th$th)
result
}
}
