% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\encoding{utf-8}
\name{sdpdth-package}
\alias{sdpdth-package}
\title{sdpdth}
\description{
M-estimator for threshold and non-threshold spatial dynamic panel data model.
}
\author{
Junyue Wu <wu.junyue.p1@dc.tohoku.ac.jp>
}
\keyword{models}
